requirements.txt
# Core dependencies (essential)
flask==2.3.3
pandas==2.1.3
openpyxl==3.1.2
flask-bootstrap==3.3.7.1
folium==0.15.0
gunicorn==21.2.0
python-dotenv==1.0.0

# Optional cloud dependencies (install only if needed)
# Uncomment these lines if using cloud data sources:
# requests==2.31.0
# boto3==1.34.0
# azure-storage-file-share==12.14.1
# google-cloud-storage==2.10.0
# google-auth==2.23.4
# google-auth-oauthlib==1.1.0

# Removed whitenoise (not needed for basic setup)
# Add back if serving large static files 